casting seconds in GSMLocation to int

+++ src/SimpleNBClientSaraR4.h
@@ -504,7 +504,7 @@ class SimpleNBSaraR4 : public SimpleNBModem<SimpleNBSaraR4>,
     lbs.year     = streamGetIntBefore(',');    // Four digit year
     lbs.hour     = streamGetIntBefore(':');    // Two digit hour
     lbs.minute   = streamGetIntBefore(':');    // Two digit minute
-    lbs.second   = streamGetFloatBefore(',');  // 6 digit second with subseconds
+    lbs.second   = static_cast<int>(streamGetFloatBefore(',')); // second only
 
     lbs.lat      = streamGetFloatBefore(',');  // latitude, in degrees
     lbs.lon      = streamGetFloatBefore(',');  // longitude, in degrees
